server/requirements.txt
Django==5.2
djangorestframework
djangorestframework-simplejwt==5.3.1
adrf==0.1.10
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.10

# Filtering and Search
django-filter==24.2

# File handling
Pillow==10.4.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.11.0

# Documentation
drf-spectacular==0.26.5

# Utilities
python-dateutil==2.8.2
pytz==2023.3


python-decouple==3.8
gunicorn==21.2.0
uvicorn[standard]==0.30.6
whitenoise==6.6.0
dj-database-url==2.0.0

# Microsoft Entra ID / Azure AD
msal==1.31.0
msgraph-sdk==1.12.0
azure-identity==1.19.0

# Data validation
pydantic==2.9.2
email-validator==2.1.0

# Development & Debugging
debugpy==1.8.0

# DocuSign Integration
docusign-esign==5.4.0
pyjwt==2.8.0
cryptography==41.0.7
